Added Lombok and updated Note entity id variable type

+++ src/main/java/com/beratkirdar/notezilla/controller/NoteController.java
@@ -24,7 +24,7 @@ public class NoteController {
     }
 
     @GetMapping("/{noteId}")
-    public Note findById(@PathVariable int noteId){
+    public Note findById(@PathVariable Long noteId){
         return noteService.findById(noteId);
     }
