+++ src/main/java/com/beratkirdar/notezilla/repository/NoteRepository.java
@@ -3,7 +3,7 @@ package com.beratkirdar.notezilla.repository;
 import com.beratkirdar.notezilla.model.Note;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoteRepository extends JpaRepository<Note, Integer> {
+public interface NoteRepository extends JpaRepository<Note, Long> {
